admin backend done

+++ Orra-Backend-SpringBoot/Orrabackend/src/main/java/com/orra/Orrabackend/dto/admin/DashboardStatsDTO.java
@@ -2,10 +2,12 @@ package com.orra.Orrabackend.dto.admin;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DashboardStatsDTO {
 
-    private double totalRevenue;
+    private BigDecimal totalRevenue;
     private long totalUsers;
     private long activeProducts;
     private long pendingApproval;
